Scheduler: fix one race condition
If 'condition_variable.notify' is called while no thread is waiting,
then the next condition_variable::wait will block.
This is why the block condition should always be checked
*while holding the lock*, before calling 'wait'.

This wasn't the case. There was a small time gap,
just after the beginning of a 'while' iteration,
and just before the call to 'condition.wait',
when one could set waitAndExit to true, and call condition.notify_one.

The notification would then be lost, and the destructor would deadlock.