Don't pass a clock by default to modules.
Only a small handful of modules actually need to deal with a clock:
renderers, and meta-level modules as rectifiers, restampers, regulators, and recorders.
Passing the system clock by default to all modules is asking for trouble.
It's safer to pass "no-clock" by default, and crash if a module tries to
access the current clock time. This forces the pipeline builder to be
explicit about what clock to use for each module.